requirements.txt
pydantic>=2.6
pyyaml>=6.0
duckdb>=1.0.0
